Restructured test file hierarchy. NO SELENIUM YET
/tests/AEGTree - now contains all AEG tests

All AEG test.ts files - references updated to include files properly

package-lock.json and package.json - Selenium installed, but not functional

vite.config.js - vitest now looks for AEG tests exclusively in /tests/AEGTree
