+++ tests/AEGTree/Ellipse.test.ts
@@ -1,8 +1,8 @@
 import {describe, expect, test} from "vitest";
 
-import {Ellipse} from "../src/AEG/Ellipse";
-import {Rectangle} from "../src/AEG/Rectangle";
-import {Point} from "../src/AEG/Point";
+import {Ellipse} from "../../src/AEG/Ellipse";
+import {Rectangle} from "../../src/AEG/Rectangle";
+import {Point} from "../../src/AEG/Point";
 
 describe("Ellipse constructor soliloquy:", () => {
     const center: Point = new Point(10, 10);
