+++ tests/AEGTree/Rectangle.test.ts
@@ -1,8 +1,8 @@
 import {describe, expect, test} from "vitest";
 
-import {Rectangle} from "../src/AEG/Rectangle";
-import {Ellipse} from "../src/AEG/Ellipse";
-import {Point} from "../src/AEG/Point";
+import {Rectangle} from "../../src/AEG/Rectangle";
+import {Ellipse} from "../../src/AEG/Ellipse";
+import {Point} from "../../src/AEG/Point";
 
 /**
  * Contains comprehensive tests on the Rectangle class.
